backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
anthropic==0.40.0
pydantic==2.9.2
dj-database-url==2.3.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
requests==2.32.3
python-decouple==3.8
pytest-django==4.9.0
pytest==8.3.3
